% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{AIC.carx}
\alias{AIC.carx}
\title{Compute the  AIC equivalent of a fitted \code{carx} object}
\usage{
\method{AIC}{carx}(object, ..., k = 2)
}
\arguments{
\item{object}{a fitted  \code{carx} object.}

\item{...}{not used.}

\item{k}{penalty for the number of parameters, default = 2.}
}
\value{
the AIC value
}
\description{
Return the AIC equivalent of a fitted \code{carx} object where the log-likelihood is replaced by the quasi-log-likelihood.
}

