% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{carx.formula}
\alias{carx.formula}
\title{provide a simple formula interface to the \code{carx} method}
\usage{
\method{carx}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{the formula.}

\item{data}{the data.}

\item{...}{other arguments supplied to \code{\link{carx.default}}.}
}
\description{
provide a simple formula interface to the \code{carx} method
}

