% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_simulate.R
\name{carx.sim}
\alias{carx.sim}
\title{simulates a sample data for \code{carx}.}
\usage{
\method{carx}{sim}(nObs, prmtrAR, prmtrX, sigmaEps, lcl, ucl, x = NULL,
  seed = NULL)
}
\arguments{
\item{nObs}{number of observations to be simulated.}

\item{prmtrAR}{the AR parameter.}

\item{prmtrX}{the parameter for X.}

\item{sigmaEps}{the standard deviation for the white noises of the AR process.}

\item{lcl}{the lower censor limit.}

\item{ucl}{the upper censor limit.}

\item{x}{optional matrix for X.}

\item{seed}{optional to set the seed of random number generator used by \code{R}.}
}
\value{
a list of simulated \code{y}, \code{x}, \code{ci}, \code{lcl} and \code{ucl}.
}
\description{
\code{simulate.carx} uses provided parameters and other settings to simulate a series of data.
}
\examples{
nObs = 100
trueX = c(0.2,0.4)
trueAR = c(-0.28,0.25)
trueSigma = 0.60
lcl = -1
ucl = 1
dat = carx.sim(100, trueAR, trueX, trueSigma, lcl, ucl, seed=0)
}

