% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_simulate.R
\name{carxSim}
\alias{carxSim}
\title{Simulate a sample data for \code{carx}}
\usage{
carxSim(nObs = 200, prmtrAR = c(-0.28, 0.25), prmtrX = c(0.2, 0.4),
  sigmaEps = 0.6, lcl = -1, ucl = 1, x = NULL, seed = NULL)
}
\arguments{
\item{nObs}{number of observations to be simulated.}

\item{prmtrAR}{the AR parameter.}

\item{prmtrX}{the parameter for X.}

\item{sigmaEps}{the standard deviation for the white noises of the AR process.}

\item{lcl}{the lower censor limit.}

\item{ucl}{the upper censor limit.}

\item{x}{optional matrix for X, default = \code{NULL}, in which case X will be simulated from standard normal distribution with dimensions determined by \code{nObs} and \code{prmtrX}.}

\item{seed}{optional to set the seed of random number generator used by \code{R}.}
}
\value{
a data frame of simulated \code{y}, \code{x}, \code{ci}, \code{lcl} and \code{ucl}.
}
\description{
Use provided parameters and other settings to simulate a series of data.
}
\examples{
dat = carxSim()
}

