% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_simulate.R
\name{carxSimCenTS}
\alias{carxSimCenTS}
\title{simulate a sample \code{\link{cenTS}} data for \code{carx}}
\usage{
carxSimCenTS(nObs = 200, prmtrAR = c(-0.28, 0.25), prmtrX = c(0.2, 0.4),
  sigmaEps = 0.6, lcl = -1, ucl = 1, x = NULL, seed = NULL,
  value.name = "y", end.date = Sys.Date())
}
\arguments{
\item{nObs}{number of observations to be simulated.}

\item{prmtrAR}{the AR parameter.}

\item{prmtrX}{the parameter for X.}

\item{sigmaEps}{the standard deviation for the white noises of the AR process.}

\item{lcl}{the lower censor limit.}

\item{ucl}{the upper censor limit.}

\item{x}{optional matrix for X, default = \code{NULL}, in which case X will be simulated from standard normal distribution with dimensions determined by \code{nObs} and \code{prmtrX}.}

\item{seed}{optional to set the seed of random number generator used by \code{R}.}

\item{value.name}{the name of the response series}

\item{end.date}{the date of the last observation, default = \code{Sys.date()}.}
}
\value{
a \code{cenTS} object with regressors.
}
\description{
Use provided parameters and other settings to simulate a series of data as a \code{cenTS} object.
}
\examples{
cts = carxSimCenTS()
}
\seealso{
\code{\link{carxSim}}.
}

