% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/carxPackage.R
\name{conditionalDistMvnorm}
\alias{conditionalDistMvnorm}
\title{calculate the conditional mean & variance of a multivariate normal distribution
with (meanVec, varMat), conditional on y at indices conditionalIndex}
\usage{
conditionalDistMvnorm(y, conditionalIndex, meanVec, varMat)
}
\arguments{
\item{y}{a vector to be conditioned on}

\item{conditionalIndex}{the index to be conditioned on}

\item{meanVec}{the mean vector the joint multivariate normal distribution}

\item{varMat}{the variance-covariance matrix of the joint multivariate normal distribution}
}
\description{
calculate the conditional mean & variance of a multivariate normal distribution
with (meanVec, varMat), conditional on y at indices conditionalIndex
}
\examples{
conditionalDistMvnorm(c(-0.5,0.5), c(2,4), c(1,2,3,4),covMat)
}
\keyword{conditional}
\keyword{distribution}
\keyword{multivariate}
\keyword{normal}

