% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_fitted.R
\name{fitted.carx}
\alias{fitted.carx}
\title{Fitted values of a \code{carx} object}
\usage{
\method{fitted}{carx}(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{not used.}
}
\value{
the fitted values.
}
\description{
Compute the fitted values from a \code{carx} object.
 Note that the existence of censoring breaks the usual Markov property for AR model.  In the presence of censoring, for each \eqn{t=1,...,n} with \eqn{n} being the sample size, we generally need to obtain the conditional distribution
 \eqn{D_t = D(Y^*_t|X_t, {(Y_{j}, X_j )}_{j=\tau}^{t-1} )},
where \eqn{1 \le \tau \le t-p} is the largest one such that none of \eqn{Y_t;t=\tau+p-1,...,\tau} is censored, due to the autoregressive nature of the regression errors. Then if \eqn{\tau = t-p}, the fitted value is computed directly as an AR process, otherwise, the fitted value is computed as the mean of the distribution \eqn{D_t} by Monte Carlo.
}

