% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_diagnostics.R
\name{goodnessOfFit.carx}
\alias{goodnessOfFit.carx}
\title{determine the goodness of fit of a fitted \code{carx} object}
\usage{
goodnessOfFit.carx(object, nLag = 10, seed = NULL, bootstrap = FALSE,
  nRep = 1000, nCores = NULL, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{nLag}{number of lags of the residuals used.}

\item{seed}{seed for random number generator.}

\item{bootstrap}{indicates whether to use the bootstrap to get the distribution of the test statistic.}

\item{nRep}{number of replications in bootstrap procedure.}

\item{nCores}{number of cores of CPU to be used in bootstrap, nCores>1 can speed up the computation. Default = \code{NULL}. See also \code{registerDoMC} in \pkg{doMC}.}

\item{...}{not used.}
}
\value{
a list of the test statistic \code{tVal} and its p-value \code{pVal}.
}
\description{
\code{goodnessOfFit.carx} determines the goodness of fit of a fitted \code{carx} model.
}

