% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_diagnostics.R
\name{lbStat}
\alias{lbStat}
\title{returns the Ljung-Box test statistic for a series with a given lag}
\usage{
lbStat(res, nLag)
}
\arguments{
\item{res}{the series.}

\item{nLag}{number of lags used.}
}
\value{
the Ljung-Box test statistic.
}
\description{
\code{lbStat} returns the Ljung-Box test statistic for a series with a given lag.
}

