% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carx_outlier.R
\name{ot2.carx}
\alias{ot2.carx}
\title{Detecting outliers in a  \code{carx} model.}
\usage{
ot2.carx(object)
}
\arguments{
\item{object}{a \code{carx} object}
}
\value{
a possibly updated \code{object} which will have an attribute \code{outlier.indices}
denoting the indices of outliers with the new index of outlier appended, if any outlier is detected.
use the probability function in tmvtnorm to test the outlier
}
\description{
This is an internal function. It tests for additional outlier one at a time, for each time point, as described in
Wang and Chan (2015), adjusted for multiplicity of testing. If the test result is significant, the function
augments the location of the most significant outlier to the vector of outlier indices,
i.e, the \code{outlier.indices} in the object returned by the function. The method used to compute the p-value is different from \code{ot.carx}.
}
\keyword{internal}
