% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phosphorusMiddleRiverIndianola.R
\name{phosphorusMiddleRiverIndianola}
\alias{phosphorusMiddleRiverIndianola}
\title{The total phosphorus concentration and river discharge data in Middle
River near Indianola, Iowa.}
\description{
This data contains the monthly total phosphorus concentration (P) and river discharge
(Q) of
the Middle River near Indianola, Iowa, USA, from 10/1999 to 10/2013. The P
data were collected  under the ambient water quality program conducted by the
Iowa Department of Natural Resources. The Q data were obtained from U.S.
Geological Survey. A gap from 09/2008 to 03/2009 in the data are due to
program suspension owing to lack
of fund. The data contains \code{logP} and \code{logQ} which are the
logarithm of the original P and Q respectively. The P data are left censored,
with -1 or 0 in \code{censorIndicator} indicating that the corresponding observation is left censored or
observed. The censoring limits are stored in \code{censorLimit}.
}
\author{
Chao Wang (chaowang@connect.hku.hk), 06/2015
}

