% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cenTS.R
\name{plot.cenTS}
\alias{plot.cenTS}
\title{Plot a \code{cenTS} object}
\usage{
\method{plot}{cenTS}(x, type = "l", auto.grid = TRUE,
  major.ticks = "auto", minor.ticks = TRUE, major.format = TRUE,
  bar.col = "grey", candle.col = "white", ann = TRUE, axes = TRUE,
  ylim = NULL, main = NULL, ...)
}
\arguments{
\item{x}{a \code{cenTS} object.}

\item{type,auto.grid,major.ticks,minor.ticks,major.format,bar.col,candle.col,ann,axes,ylim,main,...}{standard parameters to control the plot.}
}
\description{
Plot a \code{cenTS} object
}
\seealso{
\code{\link[xts]{plot.xts}}.
}

