% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_plots.R
\name{plotResAcf.carx}
\alias{plotResAcf.carx}
\title{Plotting the ACF of the residuals of a fitted \code{carx} object}
\usage{
plotResAcf.carx(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{other parameters to be supplied to \code{acf}.}
}
\value{
no value is returned, a figure will either be sent to display or be saved.
}
\description{
Plot the ACF of the residuals of a fitted \code{carx} object
}

