% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_plots.R
\name{plotResAcf}
\alias{plotResAcf}
\title{Plotting the ACF of the residuals of a fitted \code{carx} object}
\usage{
plotResAcf(object, saveFig = NULL, width = 6, height = 5, units = "in",
  res = 300, main = "", ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{saveFig}{indicates whether to save the figure, default = \code{NULL} indicates the plot will be sent to the display, otherwise, saveFig need to be a file name to save the file, with postfix ".eps", ".jpg" or ".svg".}

\item{width}{the width of the figure, if to be saved.}

\item{height}{the height of the figure, if to be saved.}

\item{units}{the units width and height of the figure.}

\item{res}{the resolution of the figure, only needed if the figure is to be saved in "jpg" format.}

\item{main}{the main title in the plot.}

\item{...}{other parameters to be supplied to \code{residuals.carx} or \code{plot}.}
}
\value{
no value is returned, a figure will either be sent to display or be saved.
}
\description{
Plot the ACF of the residuals of a fitted \code{carx} object
}

