% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_plots.R
\name{plotResiduals.carx}
\alias{plotResiduals.carx}
\title{Plotting the residual of a fitted \code{carx} object}
\usage{
plotResiduals.carx(object, residualType = "pearson", x = NULL,
  xlab = "Index", ylab = "Residual", ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{residualType}{the type of the residual.}

\item{x}{indicates x variable,
default = \code{NULL}, the time series plot of residuals will be plotted;
if x is "fitted", the residuals against the fitted values;
if x is a vector, the residuals against the supplied vector.}

\item{...}{other parameters to be supplied to \code{plot}.}
}
\description{
Plot a fitted \code{carx} object with other settings
the y axis will be the response, the x axis can be supplied by the user
}

