% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_predict.R
\name{predictARX}
\alias{predictARX}
\title{Predict a regression model with AR errors}
\usage{
predictARX(y, x, prmtrX, prmtrAR, sigma, n.ahead, newxreg, CI.level = 0.95)
}
\arguments{
\item{y}{response data used to fit the model.}

\item{x}{covariate matrix  used to fit the model.}

\item{prmtrX}{coefficients for covariates.}

\item{prmtrAR}{coefficients for AR model of the residuals.}

\item{sigma}{innovation standard deviation.}

\item{n.ahead}{The number of steps ahead to predict, default = 1.}

\item{newxreg}{The new observations for the coverates \code{x}.
If there is no covariate, the value can be assigned to be \code{NULL}.
Otherwise, a matrix of new observations is required to compute predictions.}

\item{CI.level}{The CI.level to construct the Confidence interval, default = 0.95.}
}
\description{
Predict a regression model with AR errors
}

