% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{print.carx}
\alias{print.carx}
\title{Print a short description of the fitted model}
\usage{
\method{print}{carx}(x, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{...}{not used.}
}
\value{
none.
}
\description{
Print a short description of the fitted model
}

