% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cenTS.R
\name{print.cenTS}
\alias{print.cenTS}
\title{Print a \code{cenTS} object}
\usage{
\method{print}{cenTS}(x, ...)
}
\arguments{
\item{x}{a \code{cenTS} object.}

\item{...}{not used.}
}
\value{
none.
}
\description{
Print a \code{cenTS} object
}

