% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{print.summary.carx}
\alias{print.summary.carx}
\title{print the summary of an \code{carx} object}
\usage{
\method{print}{summary.carx}(x, ...)
}
\arguments{
\item{x}{a summary of an \code{carx} object.}

\item{...}{not used.}
}
\description{
print the summary of an \code{carx} object
}

