% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_simulation.R
\name{simulation.carx}
\alias{simulation.carx}
\title{provides a simulation study for \code{carx}.}
\usage{
simulation.carx(trueX = c(0.2, 0.4), trueAR = c(-0.5, 0.3, -0.1),
  trueSigma = 0.2, lcl = -1, ucl = 1, nObs = 100, nRep = 1000,
  alpha = 0.95, nLag = 10, fullEstimation = FALSE)
}
\arguments{
\item{trueX}{the true parameter (coefficients) for exogenous variable.}

\item{trueAR}{the true AR parameter.}

\item{trueSigma}{the true sigma parameter.}

\item{lcl}{the lower censor limit.}

\item{ucl}{the upper censor limit.}

\item{nObs}{the sample size.}

\item{nRep}{the number of replications in the simulation study.}

\item{alpha}{the alpha used to construct the confidence interval for estimated parameters}

\item{nLag}{the number of lags to construct the Ljung-Box test statistic, see \code{\link{goodnessOfFit.carx}}}

\item{fullEstimation}{a bool value indicating if confidence interval by bootstrap
is performed.}
}
\value{
If nRep is 1, then the returned value is a list of the \code{carx} object and
a summary of the simulated result. If nRep > 1, then it returns a list of (nRep, objects,
averageCensorRate, alpha, summary), where
nRep is the number of replication as is supplied;
averageCensorRate is the mean of censor rates over all replications;
alpha is the alpha value for the confidence interval;
summary is a summary matrix of the simulated results.
}
\description{
\code{simulation.carx} uses provided parameters and other settings to perform a simulation study.
Note that the exogenous variables are simulated from independent standard normal.
}
\examples{
simulation.carx(nRep=1) # nRep == 1
simulation.carx(nRep=10) # nRep == 1
\dontrun{simulation.carx()} #using default settings
\dontrun{simulation.carx(ucl=Inf)} # no right censoring
\dontrun{simulation.carx(lcl=-Inf)} # no left censoring
}

