% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_plots.R
\name{tsdiag.carx}
\alias{tsdiag.carx}
\title{Show diagnostic plot for a \code{carx} object}
\usage{
\method{tsdiag}{carx}(object, gof.lag, col = "red", omit.initial = TRUE,
  mfrow = c(4, 1), main = "Diagnostic Plots", ...)
}
\arguments{
\item{object}{a \code{carx} object.}

\item{gof.lag}{the maximum number of lags in ACF and Ljung-Box goodness-of-fit test.}

\item{col}{color of some warning lines in the figures, default=\code{red}.}

\item{omit.initial}{whether initial residuals should be omitted, default = \code{TRUE}.}

\item{mfrow}{\code{par} parameter indicating how the plots are to be arranged, default = \code{c(4,1)}.}

\item{main}{The main title of the plot, default = "Diagnostic Plots".}

\item{...}{Other arguments sent to \code{plot}.}
}
\value{
none.
}
\description{
Show diagnostic plot for a \code{carx} object
}

