% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cenTS.R
\name{xreg}
\alias{xreg}
\title{Return the \code{xreg} part of the \code{cenTS} object}
\usage{
xreg(object)
}
\arguments{
\item{object}{a \code{cenTS} object.}
}
\value{
the list in \code{xreg}.
}
\description{
Return the \code{xreg} part of the \code{cenTS} object
}
\seealso{
\code{\link{cenTS}}.
}

