% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{carx.default}
\alias{carx.default}
\title{\code{carx.default} is the default method for CARX.}
\usage{
\method{carx}{default}(y, x, censorIndicator, lowerCensorLimit,
  upperCensorLimit, nAR, prmtrX = NULL, prmtrAR = NULL, sigmaEps = NULL,
  tol = 1e-04, max.iter = 500, getCI = TRUE, alpha = 0.95,
  nBootstrapSample = 1000, skipIndex = NULL, verbose = FALSE, ...)
}
\arguments{
\item{y}{a vector of possibly censored responses, only uncensored observation
marked by zeros of \code{censorIndicator} are used.}

\item{x}{a matrix of covariates, or some object which can be coerced to matrix.}

\item{censorIndicator}{a vector of -1,0,1's indicating that the corresponding y is
left-censored, not censored, and right-censored resectively.}

\item{lowerCensorLimit}{a vector of lower censor limits or a number assuming constant limit.}

\item{upperCensorLimit}{a vector of upper censor limits or a number assuming constant limit.}

\item{nAR}{the order of AR model for the regression errors.}

\item{prmtrX}{the initial value for the parameter of \code{x}, default = \code{NULL}.}

\item{prmtrAR}{the initial value for the AR coefficients, default = \code{NULL}.}

\item{sigmaEps}{the initial value for the standard deviation of the errors of the AR process, default = \code{NULL}.}

\item{tol}{the tolerance in estimating the parameters, default = 1.0e-4.}

\item{max.iter}{maximum number of iterations allowed when estimating parameters, default = 500.}

\item{getCI}{bool value to indicate if the confidence interval for the
parameter is to be constructed, default = \code{TRUE}.}

\item{alpha}{numeric value in (0,1) to get the \code{alpha} confidence interval for the parameter, default = 0.95.}

\item{nBootstrapSample}{number of bootstrap samples when estimating confidence interval for the parameter, default = 1000.}

\item{skipIndex}{a vector of indices indicating indices to be skipped, as some initial values are needed to calculate the conditional log-likelihood, also is the initial values to calculate the conditional log-likelihood, useful if there are multiple segment of series in the whole series.}

\item{verbose}{bool value indicates whether to print intermediate information, default = FALSE.}

\item{...}{not used.}
}
\value{
a CARX object of the estimated model.
}
\description{
\code{carx.default} uses given data and other settings to estimate the parameters of CARX, and optionally compute the standard error or confidence intervals of parameter estimates by parametric bootstrap.
}
\examples{
trueX = c(0.2,0.4)
nAR = 2
trueAR = c(-0.28,0.25)
trueSigma = 0.60
lcl = -1
ucl = 1
dat = simulateCarx(100, trueAR, trueX, trueSigma, lcl, ucl, seed=0)
model0 <- carx(dat$y, dat$x,censorIndicator=dat$censorIndicator,
              lowerCensorLimit = dat$lowerCensorLimit,
              upperCensorLimit = dat$upperCensorLimit,
              nAR = nAR,
              getCI = FALSE)
\dontrun{model1 <- carx(dat$y, dat$x,censorIndicator=dat$censorIndicator,
              lowerCensorLimit = dat$lowerCensorLimit,
              upperCensorLimit = dat$upperCensorLimit,
		           nAR = nAR,
		           getCI = TRUE,
              ) }
}

