% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{carx.default}
\alias{carx.default}
\title{\code{carx.default} is the default method for CARX.
\code{carx.default} uses given data and other settings to compute the estimated parameters of CARX,
and optionally compute the standard error or confidence intervals of parameter estimates
by parametric bootstrap.}
\usage{
\method{carx}{default}(y, x, censorIndicator, lowerCensorLimit,
  upperCensorLimit, nAR, tol = 1e-04, max.iter = 100, getCI = FALSE,
  alpha = 0.95, nBootstrapSample = 1000, useGoodRes = FALSE,
  skipIndex = NULL, verbose = FALSE)
}
\arguments{
\item{y}{a vector of regressors}

\item{x}{a matrix of covariates, or some object which can be coerced to matrix}

\item{censorIndicator}{a vector of -1,0,1's indicating that the corresponding y is
left-censored, not censored, or right-censored.}

\item{lowerCensorLimit}{a vector of lower censor limits for each y or a number assuming uniform limit}

\item{upperCensorLimit}{a vector of upper censor limits for each y or a number assuming uniform limit}

\item{nAR}{the order of AR model for the regression errors}

\item{tol}{the tolerance in estimating the parameters, default = 1.0e-4}

\item{max.iter}{maximum number of iterations allowed when estimating parameters, default = 500}

\item{getCI}{bool value to indicate if the confidence interval for the parameter is needed.}

\item{alpha}{numeric value (0,1) to get the confidence interval for the parameter}

\item{nBootstrapSample}{number of bootstrap samples when estimating confidence interval for the parameter, default = 1000}

\item{useGoodRes}{bool value to indicate if good values of estimated residuals are used to boostrap sample and thus estimate the confidence interval, might be useful if the normal assumption is not likely to be true, default = FALSE}

\item{skipIndex}{a vector of indices indicating indices to be skipped, as calculating the conditional log-likelihood need some initial values to start, also is the initial values to calculate the conditional log-likelihood, useful if there are multiple segment of series in the whole series.}

\item{censorLimit}{a vector of censor limits for each y}
}
\value{
a CARX object of the estimated model
}
\description{
\code{carx.default} is the default method for CARX.
\code{carx.default} uses given data and other settings to compute the estimated parameters of CARX,
and optionally compute the standard error or confidence intervals of parameter estimates
by parametric bootstrap.
}

