% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\docType{package}
\name{carx}
\alias{carx}
\alias{carx-package}
\title{\code{carx}: A package to fit Censored Auto-Regressive model with eXogenous covariates (CARX).}
\description{
\code{carx}: A package to fit Censored AutoRegressive model with eXogenous
covariates (CARX), which can also be viewed as regression models with possibly
censored responses
and autoregressive residuals. \code{carx} allows left, right, or interval
censoring for the response variable. The regression errors are assumed to
follow an autoregressive model with normal innovations. This package also
contains functions to predict the future values, diagnose whether the model is
adequate, and plot functions to illustrate the data and model.
}

