% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/carxPackage.R
\docType{package}
\name{carx}
\alias{carx}
\alias{carx-package}
\title{carx: A package for estimating censored regression models with autoregressive residuals}
\description{
carx: A package for estimating censored regression models with autoregressive residuals
}

