% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_select.R
\name{carxSelect}
\alias{carxSelect}
\title{Select the best CARX model from a variety of formulas with maximal AR order by AIC}
\usage{
carxSelect(formulas, max.ar, data = list(), detect.outlier = F, ...)
}
\arguments{
\item{formulas}{a list of model formulas, or a single formula.}

\item{max.ar}{maximal AR order.}

\item{data}{the data for CARX.}

\item{detect.outlier}{logical to determine if outlier detection is performed before the AIC for a particular model formula is computed, default = \code{FALSE}.}

\item{...}{other arguments to be supplied, if not null, it will be call with the selected model and data. Examples include {CI.compute=TRUE}, which will cause the function to re-estimate the model with confidence intervals computed, as in the selection part, the confidence interval is not computed.}
}
\value{
a list consisting of:
\itemize{
\item{\code{fitted}}{ the fitted CARX object of the model with the smallest AIC.}
\item{\code{aicMat}}{ the matrix of AIC where rows correspond to the model formulas and columns correspond to AR order.}
}
}
\description{
Select the best CARX model from a formula or a list of formulas with maximal AR order by AIC.
}
\examples{
dataSim <- carxSimCenTS(nObs=100)
fmls <- list(M1=y~X1,M2=y~X1+X2,M3=y~X1+X2-1)
\dontrun{carxSelect(y~X1,max.ar=3,data=dataSim)}
\dontrun{carxSelect(formulas=fmls,max.ar=3,data=dataSim)}
}

