% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_util.R
\name{computeCovAR}
\alias{computeCovAR}
\title{Compute the covariance matrix of some observations of the AR model}
\usage{
computeCovAR(arPrmtr, sigma, lag = length(arPrmtr) + 1)
}
\arguments{
\item{arPrmtr}{the parameter of the AR model.}

\item{sigma}{the standard deviation of the innovations of the AR model.}

\item{lag}{the number of lags to be computed, including lag at zero.}
}
\value{
the covariance matrix.
}
\description{
Compute the covariance matrix of \eqn{(\eta_t,...,\eta_{t-lag})} of the AR model.
}

