% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/carxPackage.R
\name{computeCovAR}
\alias{computeCovAR}
\title{compute the covariance matrix of {\eta_t,\cdots,\eta_{t-p}} of the AR model}
\usage{
computeCovAR(order, sigmaEps)
}
\arguments{
\item{order}{the order of the AR model}

\item{sigmaEps}{the standard deviation of the residuals of the AR model}
}
\description{
compute the covariance matrix of {\eta_t,\cdots,\eta_{t-p}} of the AR model
}

