% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/carx_util.R, R/util.R
\name{computeCovAR}
\alias{computeCovAR}
\title{compute the covariance matrix of \deqn{(\eta_t,\cdots,\eta_{t-p})} of the AR model}
\usage{
computeCovAR(order, sigmaEps)

computeCovAR(order, sigmaEps)
}
\arguments{
\item{order}{the order of the AR model}

\item{sigmaEps}{the standard deviation of the residuals of the AR model}

\item{order}{the order of the AR model}

\item{sigmaEps}{the standard deviation of the residuals of the AR model}
}
\value{
the covariance matrix needed

the covariance matrix needed
}
\description{
compute the covariance matrix of \deqn{(\eta_t,\cdots,\eta_{t-p})} of the AR model

compute the covariance matrix of \deqn{(\eta_t,\cdots,\eta_{t-p})} of the AR model
}

