% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_util.R
\name{computeCovAR}
\alias{computeCovAR}
\title{Compute the covariance matrix of some observations of the AR model}
\usage{
computeCovAR(arPrmtr, sigmaEps, lag = length(arPrmtr) + 1)
}
\arguments{
\item{arPrmtr}{the parameter of the AR model.}

\item{sigmaEps}{the standard deviation of the residuals of the AR model.}

\item{lag}{number of lags to be computed, including lag at zero.}
}
\value{
the covariance matrix needed.
}
\description{
Compute the covariance matrix of \deqn{(\eta_t,...,\eta_{t-lag})} of the AR model.
}

