% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/carxPackage.R
\name{conditionalDistMvnorm}
\alias{conditionalDistMvnorm}
\title{\code{conditionalDistMvnorm} calculates the conditional mean & variance of a random vector following multivariate normal distribution.}
\usage{
conditionalDistMvnorm(y, conditionalIndex, meanVec, varMat)
}
\arguments{
\item{y}{a vector to be conditioned on}

\item{conditionalIndex}{the index to be conditioned on}

\item{meanVec}{the mean vector the joint multivariate normal distribution}

\item{varMat}{the variance-covariance matrix of the joint multivariate normal distribution}
}
\value{
a list consisting of 'mean' and 'var' representing the conditional mean and variance respectively.
}
\description{
\code{conditionalDistMvnorm} calculates the conditional mean & variance of a sub-vector
of a vector given the rest of known elements, following multivariate normal distribution.
This function calculates the conditional mean & variance of a multivariate normal distribution.
with (meanVec, varMat), conditional on y at indices conditionalIndex.
}
\examples{
conditionalDistMvnorm(c(-0.5,0.5), c(2,4), c(1,2,3,4),matrix(
c(1,0.3,0.2,0.1, 0.3,1,-0.1,0.3,0.2,-0.1,1,0.1,0.1,0.3,0.1,1),
nrow=4,ncol=4,byrow=TRUE))
}
\keyword{conditional}
\keyword{distribution}
\keyword{multivariate}
\keyword{normal}

