% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_util.R
\name{conditionalDistMvnorm}
\alias{conditionalDistMvnorm}
\title{Calculate the conditional mean & variance of a random vector following multivariate normal distribution}
\usage{
conditionalDistMvnorm(y, conditionalIndex, meanVec, varMat)
}
\arguments{
\item{y}{the vector of the known values to be conditioned on.}

\item{conditionalIndex}{the index to be conditioned on.}

\item{meanVec}{the mean vector the joint multivariate normal distribution.}

\item{varMat}{the variance-covariance matrix of the joint multivariate normal distribution.}
}
\value{
a list consisting of 'mean' and 'var' representing the conditional mean and variance respectively.
}
\description{
Calculate the conditional mean & variance of a sub-vector
of a vector given the rest of known elements, following multivariate normal distribution.
This function calculates the conditional mean & variance of a multivariate normal distribution,
with (\code{meanVec}, \code{varMat}), conditional on y at the indices \code{conditionalIndex}.
}
\examples{
conditionalDistMvnorm(c(-0.5,0.5), c(2,4), c(1,2,3,4),matrix(
c(1,0.3,0.2,0.1, 0.3,1,-0.1,0.3,0.2,-0.1,1,0.1,0.1,0.3,0.1,1),
nrow=4,ncol=4,byrow=TRUE))
}

