% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_fitted.R
\name{fitted.carx}
\alias{fitted.carx}
\title{The fitted values from a \code{carx} object}
\usage{
\method{fitted}{carx}(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{not used.}
}
\description{
Compute the fitted values from a \code{carx} object.
}

