% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{logLik.carx}
\alias{logLik.carx}
\title{returns the log-likelihood of a \code{carx} object}
\usage{
\method{logLik}{carx}(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{not used.}
}
\value{
the expected log-likelihood.
}
\description{
returns the log-likelihood of a \code{carx} object
}

