% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carx_default.R
\name{logLik.carx}
\alias{logLik.carx}
\title{The quasi-log-likelihood of a \code{carx} object}
\usage{
\method{logLik}{carx}(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{not used.}
}
\value{
the quasi-log-likelihood value.
}
\description{
The quasi-log-likelihood of a \code{carx} object
}

