% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carx_default.R
\name{logLik.carx}
\alias{logLik.carx}
\title{The quasi-log-likelihood of a \code{carx} object}
\usage{
\method{logLik}{carx}(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{not used.}
}
\value{
the quasi-log-likelihood value.
}
\description{
The quasi-log-likelihood of a \code{carx} object
}
\examples{
dat = carxSim(nObs=100,seed=0)
mdl <- carx(y~X1+X2-1,data=dat, p=2, CI.compute = FALSE)
lk = logLik(mdl)
}

