% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_outlier.R
\name{ot.carx}
\alias{ot.carx}
\title{Detecting the outlier of the response data modelled by \code{carx}}
\usage{
ot.carx(object)
}
\arguments{
\item{object}{a \code{carx} object}
}
\value{
a \code{object} with updated information. If any outlier is detected, the return object will have a attribute \code{outlier.indices} denoting the indices of outliers.
}
\description{
Add one element to the vector of outlier indices, i.e, the \code{outlier.indices} of the object returned, if an outlier is detected.
}

