% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_outlierDetection.R
\name{ot.carx}
\alias{ot.carx}
\title{detecting the outlier of the response data modelled by \code{carx}}
\usage{
ot.carx(object)
}
\description{
detecting the outlier of the response data modelled by \code{carx}
}

