% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_plots.R
\name{plot.carx}
\alias{plot.carx}
\title{Plot a fitted \code{carx} object}
\usage{
\method{plot}{carx}(x, FUN = identity, xAxisVar = NULL, xlab = "Index",
  ylab = "Response", ...)
}
\arguments{
\item{x}{a fitted \code{carx} object.}

\item{FUN}{an optional function to be applied to the data related to the
responses of the object, e.g., if a log transformation has been applied to the data, an exp function can be supplied so that the data plotted are at the scale of the original data, default = \code{NULL}.}

\item{xAxisVar}{an optional vector to be plotted as the x variable, default =
\code{NULL} corresponds time series plot, other vectors are assumed to be
sequential.}

\item{xlab}{the label for x axis, default = "Index".}

\item{ylab}{the label for y axis, default = "Response".}

\item{...}{other parameters supplied to the generic function \code{plot}.}
}
\value{
None. A plot will be displayed.
}
\description{
\code{plot.carx} plots a fitted \code{carx} object with other settings.
The y axis will be the values related to the response.
If the object is fitted with a \code{cenTS} object which is obtained by \code{object$cenTS},
 the function will take advantage of the plot function for a \code{cenTS} object and superimpose the plot of the fitted values.
 Otherwise, the plot function will try to produce a plot similar to the previous case, while the x axis can be supplied by the user through \code{xAxisVar}, which must be ordinal and increasing.
}
\examples{
#case 1: plot with cenTS object in the object, note that the x-axis is in date.
dat = carxSimCenTS(nObs=100,seed=0)
model0 <- carx(y~X1+X2-1,data=dat, p=2, CI.compute = FALSE)
#use default settings
plot(model0)

#case 2: plot without cenTS object in the object, note that the x-axis is a vector of numbers.
dat = carxSim(nObs=100,seed=0)
model0 <- carx(y=dat$y, x=dat[,c("X1","X2")], ci=dat$ci, lcl=dat$lcl, ucl=dat$ucl, p=2)
#or simply call
model0 <- carx(y~X1+X2-1,data=dat, p=2, CI.compute = FALSE)
plot(model0)
}

