% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_plots.R
\name{plotResiduals.carx}
\alias{plotResiduals.carx}
\title{Plotting the residual of a fitted \code{carx} object}
\usage{
plotResiduals.carx(object, x = NULL, xlab = "", ylab = "",
  saveFig = NULL, width = 6, height = 5, units = "in", res = 300, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{x}{indicates x variable, default = \code{NULL}, it will plot the time series plot the residuals; if x is "fitted", it will plot the residuals against the fitted values; if x is a vector, it will plot the residual against the vector.}

\item{xlab}{the label for x axis, default = "Index".}

\item{ylab}{the label for y axis, default = "Residuals".}

\item{saveFig}{indicates whether to save the figure, default = \code{NULL} indicates the plot will be sent to the display, otherwise, saveFig need to be a file name to save the file, with postfix ".eps", ".jpg" or ".svg".}

\item{width}{the width of the figure, if to be saved.}

\item{height}{the height of the figure, if to be saved.}

\item{units}{the units width and height of the figure.}

\item{res}{the resolution of the figure, only needed if the figure is to be saved in "jpg" format.}

\item{...}{other parameters to be supplied to \code{residuals.carx} or \code{plot}.}
}
\value{
no value is returned, a figure will either be sent to display or be saved.
}
\description{
Plot a fitted \code{carx} object with other settings
the y axis will be the response, the x axis can be supplied by the user
}

