% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_predict.R
\name{predict.carx}
\alias{predict.carx}
\title{Provide predictions with fitted \code{carx} object}
\usage{
\method{predict}{carx}(object, newdata = NULL, n.ahead = 1, level = 0.95,
  nRep = 1000, ...)
}
\arguments{
\item{object}{A fitted \code{carx} object.}

\item{newdata}{The new observations for the coverates \code{x}.
If there is no covariates, the value can be assigned to be \code{NULL}.
Otherwise, an matrix of new observations is required for give predictions.}

\item{n.ahead}{The number of steps ahead the user wants to predict, default = 1.}

\item{level}{The level used to construct the Confidence interval, default = 0.95.}

\item{nRep}{The number of replications to be performed when censoring exists in the last \code{p}
observations, default = 1000.}

\item{...}{not used.}
}
\value{
A list consisting of \code{fit}, \code{se.fit}, and \code{ci} representing the predictions,
standard errors of predictions, and confidence intervals respectively.
}
\description{
\code{predict.carx} provides an method to predict the future values of an fitted
\code{carx} object with given new observations in \code{x}.
}

