% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carx_predict.R
\name{predictARX}
\alias{predictARX}
\title{Predict a regression model with AR errors}
\usage{
predictARX(y, x, prmtrX, prmtrAR, sigma, n.ahead, newxreg, CI.level = 0.95)
}
\arguments{
\item{y}{response data used to fit the model.}

\item{x}{covariate matrix  used to fit the model.}

\item{prmtrX}{regression coefficients for covariates.}

\item{prmtrAR}{AR coefficients for the regression errors.}

\item{sigma}{innovation standard deviation.}

\item{n.ahead}{the number of steps ahead to predict, default = 1.}

\item{newxreg}{the new observations for the coverates \code{x}, default={NULL}.
If there is no covariate, the value can be assigned to be \code{NULL}.
Otherwise, a matrix of new observations is required to compute predictions. Default=\code{NULL}.}

\item{CI.level}{the CI.level to construct the confidence interval, default = 0.95.}
}
\description{
Predict a regression model with AR errors
}
\keyword{internal}
