% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{print.carx}
\alias{print.carx}
\title{print a short description of the fitted model (only a few lines)}
\usage{
\method{print}{carx}(x, ...)
}
\arguments{
\item{x}{a fitted model object}
}
\description{
print a short description of the fitted model (only a few lines)
}

