% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carx_default.R
\name{print.carx}
\alias{print.carx}
\title{Print a short description of the fitted model}
\usage{
\method{print}{carx}(x, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{...}{not used.}
}
\value{
none.
}
\description{
Print a short description of the fitted model
}
\examples{
dat = carxSim(nObs=100,seed=0)
mdl <- carx(y~X1+X2-1,data=dat, p=2, CI.compute = FALSE)
print(mdl)
}

