% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carx_default.R
\name{print.summary.carx}
\alias{print.summary.carx}
\title{Print a summary of an \code{carx} object}
\usage{
\method{print}{summary.carx}(x, ...)
}
\arguments{
\item{x}{a summary of an \code{carx} object.}

\item{...}{not used.}
}
\value{
none.
}
\description{
Print a summary of an \code{carx} object
}
\examples{
dat = carxSim(nObs=100,seed=0)
mdl <- carx(y~X1+X2-1,data=dat, p=2, CI.compute = FALSE)
summary(mdl)
}
