% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/carx_residuals.R
\name{residuals.carx}
\alias{residuals.carx}
\title{get the residuals of a fitted \code{carx} object}
\usage{
\method{residuals}{carx}(object, ...)
}
\description{
get the residuals of a fitted \code{carx} object
}

