% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_residuals.R
\name{residuals.carx}
\alias{residuals.carx}
\title{compute the residuals of a fitted \code{carx} object}
\usage{
\method{residuals}{carx}(object, type = "pearson", seed = 1, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{type}{a string indicates which type of residual is to be returned.
"raw" returns the simulated residuals;
"pearson" returns the raw residuals divided by estimated standard error of the residuals.}

\item{...}{not used.}
}
\value{
the simulated residuals.
}
\description{
\code{residuals.carx} computes the simulated residuals from a fitted \code{carx} object.
}

