% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carx_default.R
\name{summary.carx}
\alias{summary.carx}
\title{summarize the fitted model on parameters, residuals and model fit}
\usage{
\method{summary}{carx}(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{not used.}
}
\value{
a summary.
}
\description{
summarize the fitted model on parameters, residuals and model fit
}

