% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carx_default.R
\name{summary.carx}
\alias{summary.carx}
\title{Summarize the fitted \code{carx} object}
\usage{
\method{summary}{carx}(object, ...)
}
\arguments{
\item{object}{a fitted \code{carx} object.}

\item{...}{not used.}
}
\value{
a summary.
}
\description{
Summarize the fitted \code{carx} object
}
\examples{
dat = carxSim(nObs=100,seed=0)
mdl <- carx(y~X1+X2-1,data=dat, p=2, CI.compute = FALSE)
summary(mdl)
}
