% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenTS.R
\name{xreg}
\alias{xreg}
\title{Return the \code{xreg} part of the \code{cenTS} object}
\usage{
xreg(object)
}
\arguments{
\item{object}{a \code{cenTS} object.}
}
\value{
the list in \code{xreg}.
}
\description{
Return the \code{xreg} part of the \code{cenTS} object
}
\examples{
strDates <- c("2000-01-01", "2000-01-02", "2000-01-03", "2000-01-04", "2000-01-05")
ts <- cenTS(value=c(1,-2,1,NA,0),
            order.by=as.Date(strDates,"\%Y-\%m-\%d"),
            lcl=c(-3,-2,-1,-1,0),
            ucl=c(3,2,1,1,1),
            x=c(1,1,1,1,1),
            y=c(2,2,2,2,2))
 xreg(ts)

}
\seealso{
\code{\link{cenTS}}.
}
